refs #629 fix use wp-env PHPUnit
