+++ tests/class-sct-phpver-judge-test.php
@@ -2,10 +2,12 @@
 
 declare( strict_types = 1 );
 
+use Yoast\WPTestUtils\WPIntegration\TestCase;
+
 /**
  * Test: Sct_Phpver_Judge
  */
-class SctPhpverJudgeTest extends PHPUnit\Framework\TestCase {
+class Sct_Phpver_Judge_Test extends TestCase {
 	/**
 	 * This test class instance.
 	 *
@@ -17,19 +19,19 @@ class SctPhpverJudgeTest extends PHPUnit\Framework\TestCase {
 	 * Settings: ABSPATH, test class file, WordPress functions.
 	 */
 	public static function setUpBeforeClass(): void {
-		if ( ! defined( 'ABSPATH' ) ) {
-			define( 'ABSPATH', '' );
+		if ( ! class_exists( 'Sct_Base ' ) ) {
+			require_once './class/class-sct-base.php';
 		}
 
 		require_once './class/class-sct-phpver-judge.php';
-		require_once './tests/lib/wordpress-functions.php';
 	}
 
 	/**
 	 * SetUp.
 	 * Create instance.
 	 */
-	protected function setUp(): void {
+	public function set_up(): void {
+		parent::set_up();
 		$this->instance = new Sct_Phpver_Judge();
 	}
